Move dashboard authorization checks into their own class
