+++ app/authorization/policies/dashboard_policy.rb
@@ -1,0 +1,12 @@
+class DashboardPolicy < ApplicationPolicy
+
+  def admin_dashboard?
+    is_admin?
+  end
+
+  def member_dashboard?
+    is_member?
+  end
+
+end
+
