+++ app/authorization/policies/user_policy.rb
@@ -10,14 +10,6 @@ class UserPolicy < ApplicationPolicy
     end
   end
 
-  def admin_dashboard?
-    is_admin?
-  end
-
-  def member_dashboard?
-    is_member?
-  end
-
 # CRUD
 
   # Admins can't delete themselves.
