+++ app/controllers/admin/dashboard_controller.rb
@@ -1,7 +1,7 @@
 class Admin::DashboardController < Admin::BaseController
 
   def index
-    authorize(User, :admin_dashboard?)
+    authorize(:dashboard, :admin_dashboard?)
   end
 
 end
